requirements.txt
anthropic==0.86.0
apscheduler==3.11.2
flask==3.1.3
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
groq==1.1.1
httpx==0.28.1
python-dotenv==1.2.2
sentry-sdk==2.56.0
slack-bolt==1.27.0
